rework of Edges
Added:
- function `new` to take `Edges` from any data.
- `BinImage` structure to represent an image.
- `Point` alias to represent a point in the image.
Moved:
- structure `Edges` moved to lib.rs.
- tests to separate file.
- `distance` function to utils.
Removed:
- `hashbrown`, `mashmap` and `thiserror` from dependencies.
